Refactor: Implement protected layout pattern for authenticated routes
- Create (protected) route group with global auth middleware
  - Layout handles authentication check and redirect to /login for unauthenticated users
  - Includes Sidebar component for all protected pages
  - Clean separation from public routes

- Move protected routes into (protected) group:
  - /favorites
  - /library
  - /settings
  - /catalogue
  - /search (redirects to /catalogue)
  - /plan
  - /logout

- Create (public) route group for public pages:
  - /login
  - /register

- Simplify page components:
  - Remove duplicate auth checks from individual pages
  - Remove Sidebar imports from pages (now in layout)
  - Pages now focus on content only

- Update FavoritePage:
  - Remove router-based auth handling
  - Simplify useEffect to only track userId
  - Delete authChecked and isLoggedIn state (now in layout)
