+++ src/app/[lang]/(protected)/favorites/page.tsx
@@ -1,0 +1,11 @@
+'use client';
+
+import type { ReactElement } from 'react';
+
+import FavoritePage from '@/components/templates/FavoritePage';
+
+const Favorites = (): ReactElement => {
+  return <FavoritePage />;
+};
+
+export default Favorites;
