+++ src/app/[lang]/(protected)/layout.tsx
@@ -1,0 +1,48 @@
+'use client';
+
+import { onAuthStateChanged } from 'firebase/auth';
+import { useRouter } from 'next/navigation';
+import { useEffect, useState, type ReactNode } from 'react';
+
+import Sidebar from '@/organisms/Sidebar';
+import { auth } from '@/config/firebase';
+
+interface ProtectedLayoutProps {
+  children: ReactNode;
+}
+
+export default function ProtectedLayout({ children }: ProtectedLayoutProps) {
+  const router = useRouter();
+  const [authChecked, setAuthChecked] = useState(false);
+  const [isLoggedIn, setIsLoggedIn] = useState(false);
+
+  useEffect(() => {
+    const unsubscribe = onAuthStateChanged(auth, user => {
+      if (user) {
+        setIsLoggedIn(true);
+      } else {
+        setIsLoggedIn(false);
+      }
+      setAuthChecked(true);
+    });
+
+    return () => unsubscribe();
+  }, []);
+
+  useEffect(() => {
+    if (authChecked && !isLoggedIn) {
+      router.replace('/login');
+    }
+  }, [authChecked, isLoggedIn, router]);
+
+  if (!authChecked || !isLoggedIn) {
+    return null;
+  }
+
+  return (
+    <>
+      <Sidebar />
+      {children}
+    </>
+  );
+}
