+++ src/app/[lang]/(protected)/logout/page.tsx
@@ -1,0 +1,9 @@
+'use client';
+
+import type { ReactElement } from 'react';
+
+import LogoutPage from '@/components/templates/LogoutPage';
+
+const Logout = (): ReactElement => <LogoutPage />;
+
+export default Logout;
