+++ src/app/[lang]/(protected)/search/page.tsx
