+++ src/app/[lang]/(public)/register/page.tsx
