--- src/app/[lang]/logout/page.tsx
@@ -1,14 +1,0 @@
-'use client';
-import { type ReactElement } from 'react';
-
-import LogoutPage from '@/components/templates/LogoutPage';
-import Sidebar from '@/organisms/Sidebar';
-
-const Login = (): ReactElement => (
-  <>
-    <Sidebar />
-    <LogoutPage />
-  </>
-);
-
-export default Login;
